Refactor task creation view
