+++ src/pages/ManageTasksVariants.vue
@@ -26,8 +26,8 @@
 <script setup>
 import PvTabPanel from 'primevue/tabpanel';
 import PvTabView from 'primevue/tabview';
-import ManageTasks from '../components/tasks/ManageTasks.vue';
-import ManageVariants from '../components/tasks/ManageVariants.vue';
+import ManageTasks from '@/containers/ManageTasks';
+import ManageVariants from '@/components/tasks/ManageVariants.vue';
 </script>
 
 <style scoped>
